docs: fix hydration (#39153)
* docs: fix hydration

* chore: recover style
